Refactor tangle events (#952)
* Feat: added booking of data messages

* Feat: added ApproverType

* Fix: fixed prefix iteration for Approvers

* Fix: temporarily changed TransactionType of payload package

* Feat: added optional ApproverType for Approvers

* Refactor: renamed a method

* Feat: added IsInPastCone (not fully done, yet)

* Feat: added pastconecheck for messages

* Feat: WIPWIPWIPWIP

* WIP2WIP2WIP2

* Fix: execute tests online

* Fix: fixed test

* Fix: fixed some stuff

* Fix: fixed invalid bug

* Fix: fixed test issues on windows

* Fix: fixed last issues with the test

* Refactor: refactored remaining code

* Fix: fixed bug with tipselector

* Fix: fixed solidification

* Refactor: removed extraneous Solidification

* Fix: fixed walker

* Docs: fixed the comment to match the refactor

* Docs: changed docs again

* Feat: Parser initialized in tangle

* Fix: fixed some issues

* Feat: refactored some code

* Refactor: changed code style to guards

* Docs: fixed missing comment

* Refactor: consolidated parents age check

* Docs: fixed missing article in comment

* Refactor: cleaned up code

* Refactor: last refactor

* Feat: moved Walk methods to Utils

* Refactor: started implementing generic walker

* Feat: moved Walker to hive.go

* Refactor: refactored code

* Fix: fixed merge bug

* Refactor: changed test according to refactor

* Feat: updated hive.go to latest

* Feat: finished implementing the Booker

* Feat: started implementing Options + refactored Solidifer a bit

* Refactor: moved MessageFactory initialization to tangle

* Feat: added first Tangle options

* Refactor: changed something

* 🐛 Fix TipSelector initialization

* ✅ Remove old and stupid unit test

* ♻️ Adds new constructor

* Feat: moved all initialization tasks to Tangle

* Refactor: fixed erroneous refactor

* Fix: fixed refactor issues

* Refactor: added logging to messagelayer plugin

* Feat: added a MarkerIndexBranchIDMapping data structure

* Feat: added a mapping between markers and branchIDs

* Refactor: refactored some code

* Fix: fixed missing refactor

* Fix: fixed some refactor issues

* Feat: started adding Setup routines for the dataflow

* Feat: started adding more Setup routines

* Refactor: refactored some more stuff :P

* Perf: further refactor

* Fix: fixed CachedMessageEvent

* Refactor: remove CachedMessageEvent

* Fix: fixed last bugs after refactor

* Docs: refactored some docs

* Refactor: restructured files

* Fix parser test fail

* Refactor: moved events to structs
