--- packages/tangle/events.go
@@ -1,107 +1,0 @@
-package tangle
-
-import (
-	"github.com/iotaledger/hive.go/autopeering/peer"
-	"github.com/iotaledger/hive.go/events"
-)
-
-// MessageTipSelectorEvents represents event happening on the tip-selector.
-type MessageTipSelectorEvents struct {
-	// Fired when a tip is added.
-	TipAdded *events.Event
-	// Fired when a tip is removed.
-	TipRemoved *events.Event
-}
-
-func newMessageTipSelectorEvents() *MessageTipSelectorEvents {
-	return &MessageTipSelectorEvents{
-		TipAdded:   events.NewEvent(messageIDEventHandler),
-		TipRemoved: events.NewEvent(messageIDEventHandler),
-	}
-}
-
-// MessageFactoryEvents represents events happening on a message factory.
-type MessageFactoryEvents struct {
-	// Fired when a message is built including tips, sequence number and other metadata.
-	MessageConstructed *events.Event
-	// Fired when an error occurred.
-	Error *events.Event
-}
-
-func newMessageFactoryEvents() *MessageFactoryEvents {
-	return &MessageFactoryEvents{
-		MessageConstructed: events.NewEvent(messageConstructedEvent),
-		Error:              events.NewEvent(events.ErrorCaller),
-	}
-}
-
-// MessageParserEvents represents events happening on a message parser.
-type MessageParserEvents struct {
-	// Fired when a message was parsed.
-	MessageParsed *events.Event
-	// Fired when submitted bytes are rejected by a filter.
-	BytesRejected *events.Event
-	// Fired when a message got rejected by a filter.
-	MessageRejected *events.Event
-}
-
-// MessageParsedEvent represents the parameters of messageParsedEvent
-type MessageParsedEvent struct {
-	Message *Message
-	Peer    *peer.Peer
-}
-
-// BytesRejectedEvent represents the parameters of bytesRejectedEvent
-type BytesRejectedEvent struct {
-	Bytes []byte
-	Peer  *peer.Peer
-}
-
-// MessageRejectedEvent represents the parameters of messageRejectedEvent
-type MessageRejectedEvent struct {
-	Message *Message
-	Peer    *peer.Peer
-}
-
-func newMessageParserEvents() *MessageParserEvents {
-	return &MessageParserEvents{
-		MessageParsed:   events.NewEvent(messageParsedEvent),
-		BytesRejected:   events.NewEvent(bytesRejectedEvent),
-		MessageRejected: events.NewEvent(messageRejectedEvent),
-	}
-}
-
-// SendRequestEvent represents the parameters of sendRequestEvent
-type SendRequestEvent struct {
-	ID MessageID
-}
-
-func newMessageRequesterEvents() *MessageRequesterEvents {
-	return &MessageRequesterEvents{
-		SendRequest: events.NewEvent(sendRequestEvent),
-	}
-}
-
-func sendRequestEvent(handler interface{}, params ...interface{}) {
-	handler.(func(*SendRequestEvent))(params[0].(*SendRequestEvent))
-}
-
-func messageParsedEvent(handler interface{}, params ...interface{}) {
-	handler.(func(*MessageParsedEvent))(params[0].(*MessageParsedEvent))
-}
-
-func bytesRejectedEvent(handler interface{}, params ...interface{}) {
-	handler.(func(*BytesRejectedEvent, error))(params[0].(*BytesRejectedEvent), params[1].(error))
-}
-
-func messageRejectedEvent(handler interface{}, params ...interface{}) {
-	handler.(func(*MessageRejectedEvent, error))(params[0].(*MessageRejectedEvent), params[1].(error))
-}
-
-func messageConstructedEvent(handler interface{}, params ...interface{}) {
-	handler.(func(*Message))(params[0].(*Message))
-}
-
-func messageIDEventHandler(handler interface{}, params ...interface{}) {
-	handler.(func(MessageID))(params[0].(MessageID))
-}
