+++ packages/tangle/tangle.go
@@ -101,6 +101,10 @@ type Events struct {
 	Error *events.Event
 }
 
+func messageIDEventHandler(handler interface{}, params ...interface{}) {
+	handler.(func(MessageID))(params[0].(MessageID))
+}
+
 // endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 // region Options //////////////////////////////////////////////////////////////////////////////////////////////////////
