+++ packages/tangle/tipmanager.go
@@ -5,11 +5,13 @@ import (
 	"github.com/iotaledger/hive.go/events"
 )
 
+// region TipManager ///////////////////////////////////////////////////////////////////////////////////////////////////
+
 // TipManager manages a map of tips and emits events for their removal and addition.
 type TipManager struct {
 	tangle *Tangle
 	tips   *randommap.RandomMap
-	Events *MessageTipSelectorEvents
+	Events *TipManagerEvents
 }
 
 // NewTipManager creates a new tip-selector.
@@ -17,7 +19,10 @@ func NewTipManager(tangle *Tangle, tips ...MessageID) *TipManager {
 	tipSelector := &TipManager{
 		tangle: tangle,
 		tips:   randommap.New(),
-		Events: newMessageTipSelectorEvents(),
+		Events: &TipManagerEvents{
+			TipAdded:   events.NewEvent(messageIDEventHandler),
+			TipRemoved: events.NewEvent(messageIDEventHandler),
+		},
 	}
 
 	if tips != nil {
@@ -88,3 +93,18 @@ func (t *TipManager) Tips(count int) (parents []MessageID) {
 func (t *TipManager) TipCount() int {
 	return t.tips.Size()
 }
+
+// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+// region TipManagerEvents /////////////////////////////////////////////////////////////////////////////////////////////
+
+// TipManagerEvents represents events happening on the TipManager.
+type TipManagerEvents struct {
+	// Fired when a tip is added.
+	TipAdded *events.Event
+
+	// Fired when a tip is removed.
+	TipRemoved *events.Event
+}
+
+// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
